# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reel Journal API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ConnectInstagramBody {
  accessToken: string;
}

export interface ConnectInstagramResponse {
  success: boolean;
  username: string;
  accountId: string;
}

export interface InstagramStatus {
  connected: boolean;
  /** @nullable */
  username: string | null;
  /** @nullable */
  accountId: string | null;
  /** @nullable */
  lastSynced: string | null;
}

export interface SyncResult {
  synced: number;
  total: number;
  message: string;
}

export interface HashtagMedia {
  id: string;
  mediaType: string;
  /** @nullable */
  permalink?: string | null;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  commentsCount?: number | null;
  /** @nullable */
  likeCount?: number | null;
  /** @nullable */
  timestamp?: string | null;
}

export interface HashtagSearchResult {
  hashtag: string;
  media: HashtagMedia[];
}

export interface Reel {
  id: number;
  instagramId: string;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  permalink?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  mediaUrl?: string | null;
  /** @nullable */
  postedAt?: string | null;
  /** @nullable */
  likeCount?: number | null;
  /** @nullable */
  commentsCount?: number | null;
  /** @nullable */
  reach?: number | null;
  /** @nullable */
  saves?: number | null;
  /** @nullable */
  shares?: number | null;
  /** @nullable */
  plays?: number | null;
  /**
   * underperforming | normal | overperforming
   * @nullable
   */
  performanceStatus?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface ReelNotes {
  id: number;
  reelId: number;
  /** @nullable */
  hook?: string | null;
  /** @nullable */
  format?: string | null;
  /** @nullable */
  ideaSource?: string | null;
  /** @nullable */
  whyItWorked?: string | null;
  /** @nullable */
  whyItFailed?: string | null;
  /** @nullable */
  emotionalReaction?: string | null;
  /**
   * funny | controversial | relatable | weird | cinematic | motivational
   * @nullable
   */
  contentType?: string | null;
  /** @nullable */
  wouldRemake?: boolean | null;
  /** @nullable */
  inspirationLink?: string | null;
  /** @nullable */
  extraNotes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ReelAnalysis {
  id: number;
  reelId: number;
  summary: string;
  performanceDrivers: string;
  retentionFactors: string;
  contentPatterns: string;
  lessonsLearned: string;
  nextIdeas: string;
  variablesToRepeat: string;
  createdAt: string;
  updatedAt: string;
}

export type ReelDetail = Reel & {
  notes?: ReelNotes;
  analysis?: ReelAnalysis;
};

export interface ReelAverages {
  avgLikes: number;
  avgComments: number;
  /** @nullable */
  avgReach: number | null;
  /** @nullable */
  avgSaves: number | null;
  /** @nullable */
  avgShares: number | null;
}

export interface DashboardSummary {
  latestReel?: ReelDetail;
  totalReels: number;
  averages: ReelAverages;
  connected: boolean;
  /** @nullable */
  lastSynced: string | null;
}

export interface ListReelsResponse {
  reels: Reel[];
  total: number;
}

export interface SaveReelNotesBody {
  /** @nullable */
  hook?: string | null;
  /** @nullable */
  format?: string | null;
  /** @nullable */
  ideaSource?: string | null;
  /** @nullable */
  whyItWorked?: string | null;
  /** @nullable */
  whyItFailed?: string | null;
  /** @nullable */
  emotionalReaction?: string | null;
  /** @nullable */
  contentType?: string | null;
  /** @nullable */
  wouldRemake?: boolean | null;
  /** @nullable */
  inspirationLink?: string | null;
  /** @nullable */
  extraNotes?: string | null;
}

export interface UpdateReelTagsBody {
  tags: string[];
}

export interface UpdateReelTagsResponse {
  tags: string[];
}

export interface PlaybookLesson {
  id: number;
  lesson: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  sourceReelId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface ListPlaybookResponse {
  lessons: PlaybookLesson[];
}

export interface CreatePlaybookBody {
  lesson: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  sourceReelId?: number | null;
}

export interface UpdatePlaybookBody {
  lesson?: string;
  /** @nullable */
  category?: string | null;
}

export interface SavedReference {
  id: number;
  url: string;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  whyItsgood?: string | null;
  /** @nullable */
  whatToChange?: string | null;
  /** @nullable */
  howToRemake?: string | null;
  /** @nullable */
  commentsCount?: number | null;
  /** @nullable */
  likeCount?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface ListReferencesResponse {
  references: SavedReference[];
}

export interface CreateReferenceBody {
  url: string;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  whyItsgood?: string | null;
  /** @nullable */
  whatToChange?: string | null;
  /** @nullable */
  howToRemake?: string | null;
  /** @nullable */
  commentsCount?: number | null;
  /** @nullable */
  likeCount?: number | null;
}

export interface UpdateReferenceBody {
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  whyItsgood?: string | null;
  /** @nullable */
  whatToChange?: string | null;
  /** @nullable */
  howToRemake?: string | null;
  /** @nullable */
  commentsCount?: number | null;
  /** @nullable */
  likeCount?: number | null;
}

export type SearchHashtagParams = {
  hashtag: string;
  limit?: number;
};

export type ListReelsParams = {
  sortBy?: ListReelsSortBy;
  sortOrder?: ListReelsSortOrder;
  /**
   * Comma-separated tags to filter by
   */
  tags?: string;
  limit?: number;
  offset?: number;
};

export type ListReelsSortBy =
  (typeof ListReelsSortBy)[keyof typeof ListReelsSortBy];

export const ListReelsSortBy = {
  createdAt: "createdAt",
  likeCount: "likeCount",
  commentsCount: "commentsCount",
  reach: "reach",
  saves: "saves",
  shares: "shares",
} as const;

export type ListReelsSortOrder =
  (typeof ListReelsSortOrder)[keyof typeof ListReelsSortOrder];

export const ListReelsSortOrder = {
  asc: "asc",
  desc: "desc",
} as const;