# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Reel Journal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /instagram/connect:
    post:
      operationId: connectInstagram
      tags: [instagram]
      summary: Save Instagram access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectInstagramBody"
      responses:
        "200":
          description: Token saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectInstagramResponse"
        "400":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /instagram/sync:
    post:
      operationId: syncReels
      tags: [instagram]
      summary: Fetch and sync latest Reels from Instagram
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
        "400":
          description: Not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /instagram/hashtag-search:
    get:
      operationId: searchHashtag
      tags: [instagram]
      summary: Search for media by hashtag via Instagram API
      parameters:
        - name: hashtag
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Hashtag media results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HashtagSearchResult"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── DASHBOARD ──────────────────────────────────────────────────────────────
...
  /reels/{id}/analyze:
    post:
      operationId: analyzeReel
      tags: [analysis]
      summary: Run AI analysis on a Reel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReelAnalysis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Analysis failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reels/{id}/analysis:
    get:
      operationId: getReelAnalysis
      tags: [analysis]
      summary: Get stored AI analysis for a Reel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stored analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReelAnalysis"
        "404":
          description: No analysis yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── PLAYBOOK ────────────────────────────────────────────────────────────────
  /playbook:
    get:
      operationId: listPlaybookLessons
      tags: [playbook]
      summary: List all playbook lessons
      responses:
        "200":
          description: Playbook lessons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListPlaybookResponse"
    post:
      operationId: createPlaybookLesson
      tags: [playbook]
      summary: Create a new playbook lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlaybookBody"
      responses:
        "201":
          description: Created lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybookLesson"
  /playbook/{id}:
    patch:
      operationId: updatePlaybookLesson
      tags: [playbook]
      summary: Update a playbook lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlaybookBody"
      responses:
        "200":
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybookLesson"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePlaybookLesson
      tags: [playbook]
      summary: Delete a playbook lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SAVED REFERENCES (Remake List) ─────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── INSTAGRAM ────────────────────────────────────────────────────────────
    ConnectInstagramBody:
      type: object
      properties:
        accessToken:
          type: string
      required:
        - accessToken
    ConnectInstagramResponse:
      type: object
      properties:
        success:
          type: boolean
        username:
          type: string
        accountId:
          type: string
      required:
        - success
        - username
        - accountId
...
    SyncResult:
      type: object
      properties:
        synced:
          type: integer
        total:
          type: integer
        message:
          type: string
      required:
        - synced
        - total
        - message
    HashtagSearchResult:
      type: object
      properties:
        hashtag:
          type: string
        media:
          type: array
          items:
            $ref: "#/components/schemas/HashtagMedia"
      required:
        - hashtag
        - media
    HashtagMedia:
      type: object
      properties:
        id:
          type: string
        mediaType:
          type: string
        permalink:
          type: ["string", "null"]
        caption:
          type: ["string", "null"]
        commentsCount:
          type: ["integer", "null"]
        likeCount:
          type: ["integer", "null"]
        timestamp:
          type: ["string", "null"]
      required:
        - id
        - mediaType

    # ─── DASHBOARD ────────────────────────────────────────────────────────────
...
    ReelAnalysis:
      type: object
      properties:
        id:
          type: integer
        reelId:
          type: integer
        summary:
          type: string
        performanceDrivers:
          type: string
        retentionFactors:
          type: string
        contentPatterns:
          type: string
        lessonsLearned:
          type: string
        nextIdeas:
          type: string
        variablesToRepeat:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - reelId
        - summary
        - performanceDrivers
        - retentionFactors
        - contentPatterns
        - lessonsLearned
        - nextIdeas
        - variablesToRepeat
        - createdAt
        - updatedAt

    # ─── PLAYBOOK ─────────────────────────────────────────────────────────────
    PlaybookLesson:
      type: object
      properties:
        id:
          type: integer
        lesson:
          type: string
        category:
          type: ["string", "null"]
        sourceReelId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - lesson
        - createdAt
        - updatedAt
    ListPlaybookResponse:
      type: object
      properties:
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/PlaybookLesson"
      required:
        - lessons
    CreatePlaybookBody:
      type: object
      properties:
        lesson:
          type: string
        category:
          type: ["string", "null"]
        sourceReelId:
          type: ["integer", "null"]
      required:
        - lesson
    UpdatePlaybookBody:
      type: object
      properties:
        lesson:
          type: string
        category:
          type: ["string", "null"]

    # ─── SAVED REFERENCES ─────────────────────────────────────────────────────
...