# lib/api-zod/src/generated/types/reelAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reel Journal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReelAnalysis {
  id: number;
  reelId: number;
  summary: string;
  performanceDrivers: string;
  retentionFactors: string;
  contentPatterns: string;
  lessonsLearned: string;
  nextIdeas: string;
  variablesToRepeat: string;
  createdAt: string;
  updatedAt: string;
}