# lib/api-zod/src/generated/types/updatePlaybookBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reel Journal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePlaybookBody {
  lesson?: string;
  /** @nullable */
  category?: string | null;
}